+++ resources/views/components/slideover/slideover.blade.php
@@ -74,7 +74,7 @@ Nested slideovers:
     $closeId = $isInForm ? 'close-' . $id : $id;
 @endphp
 
-<x-rapidez::tag :is="$tag" class="relative z-slideover">
+<x-rapidez::tag :is="$tag">
     <input id="{{ 'close-' . $id }}" class="hidden" type="reset">
     @if (!$hasParent)
         <input @checked($open) id="{{ $id }}" class="peer hidden prevent-scroll" type="checkbox">