+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog 
 
-[Unreleased changes](https://github.com/rapidez/blade-components/compare/1.0.2...1.0.2)
+[Unreleased changes](https://github.com/rapidez/blade-components/compare/1.3.0...1.3.0)
+## [1.3.0](https://github.com/rapidez/blade-components/releases/tag/1.3.0) - 2025-01-09
+
+### Added
+
+- Readmore component (#5)
+
+## [1.2.0](https://github.com/rapidez/blade-components/releases/tag/1.2.0) - 2024-12-04
+
+### Added
+
+- Slideover component (#10)
+
+## [1.1.0](https://github.com/rapidez/blade-components/releases/tag/1.1.0) - 2024-12-03
+
+### Added
+
+- Prose component (#11, 674a8ac)
+
 ## [1.0.2](https://github.com/rapidez/blade-components/releases/tag/1.0.2) - 2024-11-27
 
 ### Fixed